pyproject.toml
[tool.poetry]
name = "squirrelfocus"
version = "0.1.0"
description = "Developer productivity and reflection toolkit."
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.3"
openai = "^1.14.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.3.0"
ruff = "^0.4.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
sf = "cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ["py310"]
